refactor code - divide class architecture
